Modified CROCO diagnosis script written in MATLAB
This script is dedicated to diagnosing (i.e., determining the spin-up time) of a climatological CROCO simulation. Place this script in your croco_tools/Diagnostic_tools directory. The script is modified from the original croco_diags.m that comes with the CROCO tools. Note that, the original croco_diags.m is designed for diagnosing inter-annual simulations.